refactor: restructure api
