+++ docker-compose.yaml
@@ -17,11 +17,15 @@ services:
       - postgres
   postgres:
     image: postgres:16
+    ports:
+      - "5432:5432"
     environment:
       - POSTGRES_USER=postgres
       - POSTGRES_PASSWORD=postgres
       - POSTGRES_DB=postgres
     volumes:
       - postgres_data:/var/lib/postgresql/data
+      - ./api/app/data/init:/docker-entrypoint-initdb.d
+
 volumes:
   postgres_data:
