+++ shared/opsbox_common/models.py
@@ -17,9 +17,10 @@ class TaskStatus(enum.Enum):
 
 
 class Task(Base):
-    __tablename__ = "task"
+    __tablename__ = "tasks"
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
     title = Column(String(255), index=True)
     status = Column(Enum(TaskStatus, name="task_status"), default=TaskStatus.NEW, nullable=False)
     created_at = Column(DateTime, default=datetime.utcnow, nullable=False)
+    updated_at = Column(DateTime, default=datetime.utcnow, onupdate=datetime.utcnow, nullable=False)
     result = Column(String(255), default=None, nullable=True)
